+++ slack-gamebot/service.rb
@@ -1,4 +1,7 @@
 module SlackGamebot
   class Service < SlackRubyBotServer::Service
+    def self.instance
+      SlackRubyBotServer::Service.instance(server_class: SlackGamebot::Server)
+    end
   end
 end